+++ hosts/boomer-kuwanger/default.nix
@@ -36,7 +36,6 @@
     frontend = "es-de";
     theme = "art-book-next";
     esde.appDataDir = "/srv/emulation/es-de";
-    visuals.defaultProfile = "nnedi3-clean";
     visuals.upscaler = "none";
     controllers.assignment = "connection-order-persistent";
     startup.mode = "kiosk";